nahuatl_translator_project/requirements.txt
# Pinned to avoid common Colab/Python3.12 issues
numpy>=2.0,<2.3
pandas>=2.2
openpyxl>=3.1
huggingface-hub>=0.34.0,<2.0
safetensors==0.4.5
tokenizers==0.20.3
transformers==4.46.3
accelerate==1.1.1
datasets==2.21.0
sentencepiece==0.2.0
sacrebleu==2.4.2
torch

# Clean UI + uploads
gradio==4.44.0

# Document parsing / OCR
pymupdf>=1.24.0
pytesseract>=0.3.10
pillow>=10.0.0

# Optional: higher-quality translation + structured extraction
openai>=1.40.0

# Dev/test
pytest>=8.0.0
